refactor: header stats
-Refactor profile header stats so that the layout doesn't overflow on smaller screens.
